adder stolpe til draw + dynamisk viewbox

+++ src/main/java/business/services/SVG.java
@@ -20,7 +20,7 @@ public class SVG {
             "y=\"%d\"   " +
             " preserveAspectRatio=\"xMinYMin\">";
 
-    private final String rectTemplate = "<rect x=\"%d\" y=\"%d\" height=\"%f\" width=\"%f\" style=\"stroke:#000000; fill: #ffffff\" />";
+    private final String rectTemplate = "<rect x=\"%f\" y=\"%f\" height=\"%f\" width=\"%f\" style=\"stroke:#000000; fill: #ffffff\" />";
 
     public SVG(int x, int y, String viewBox, int width, int height) {
         this.x = x;
@@ -31,7 +31,7 @@ public class SVG {
         svg.append(String.format(headerTemplate, height, width, viewBox, x, y));
     }
 
-    public void addRect(int x, int y, double height, double width) {
+    public void addRect(double x, double y, double height, double width) {
         svg.append(String.format(Locale.US, rectTemplate, x, y, height, width));
     }
